fix host tab-fetchNetworks
